feat: Componentize tweet #118
